refactor: category dialogs
